refactor: create task-usecase, useTasks
