+++ src/app/App.tsx
@@ -1,13 +1,10 @@
 import { RoconRoot } from "rocon/react"
-import { YarukotoContextProvider } from "../context/YarukotoContext"
 import { Routes } from "./Router"
 
 export const App = () => {
   return (
-    <YarukotoContextProvider>
-      <RoconRoot>
-        <Routes />
-      </RoconRoot>
-    </YarukotoContextProvider>
+    <RoconRoot>
+      <Routes />
+    </RoconRoot>
   )
 }
